+++ app/src/main/java/woowacourse/shopping/data/model/CartItemEntity.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.model
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import woowacourse.shopping.data.db.cartItem.CartItemDatabase.Companion.CART_ITEMS_DB_NAME
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.Product
 
 @Entity(tableName = CART_ITEMS_DB_NAME)
 data class CartItemEntity(
