+++ app/src/main/java/woowacourse/shopping/domain/model/CartItem.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 data class CartItem(val id: Long, val product: Product)
