+++ app/src/main/java/woowacourse/shopping/domain/model/Product.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 data class Product(val id: Long, val name: String, val price: Int, val imageUrl: String)
