+++ app/src/main/java/woowacourse/shopping/view/products/adapter/ProductAdapter.kt
@@ -5,7 +5,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemProductBinding
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.view.products.OnClickProduct
 import woowacourse.shopping.view.products.adapter.viewholder.ProductViewHolder
