+++ app/src/main/java/woowacourse/shopping/view/products/adapter/viewholder/ProductViewHolder.kt
@@ -4,7 +4,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.target.Target
 import woowacourse.shopping.databinding.ItemProductBinding
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.view.products.OnClickProduct
 
 class ProductViewHolder(
