+++ app/src/test/java/woowacourse/shopping/domain/ShoppingCartTest.kt
@@ -3,6 +3,9 @@ package woowacourse.shopping.domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import woowacourse.shopping.domain.model.CartItem
+import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.model.ShoppingCart
 
 class ShoppingCartTest {
     private lateinit var shoppingCart: ShoppingCart
